[rs-storage] Update delete process to handle more files.

+++ rs-storage/storage/storage-service/src/main/java/fr/cnes/regards/modules/storage/service/IAIPService.java
@@ -236,7 +236,7 @@ public interface IAIPService {
      * Schedule deletion of datafiles marked for deletion
      * @return number of scheduled aip to delete
      */
-    int doDelete();
+    Long doDelete();
 
     /**
      * Remove {@link AIP}s associated the given sip, through its ip id
